+++ Gemfile
@@ -7,6 +7,7 @@ gemspec
 gem "activesupport", "6.1.4.1"
 gem "foreman", "0.87.2"
 gem "google-cloud-pubsub", "2.8.1"
+gem "graphql", "1.12.19"
 gem "paint", "2.2.1"
 gem "parser", "3.0.2.0"
 gem "pg", "1.2.3"
