+++ lib/souls/app/index.rb
@@ -1,0 +1,4 @@
+require_relative "graphql/souls_query"
+
+module Souls
+end
