+++ lib/souls/index.rb
@@ -1,7 +1,9 @@
 require_relative "./version"
 require "thor"
+require "graphql"
 require_relative "./cli/index"
 require_relative "./utils/index"
+require_relative "./app/index"
 
 module Souls
 end
